+++ aqueduct/shm/__init__.py
@@ -1,0 +1,3 @@
+from .mixin import SharedFieldsMixin
+
+__all__ = ['SharedFieldsMixin']