+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Unreleased
-- Fix setup requirements
+- Update setup requirements: add extras - aiohttp, numpy
 
 # 1.11.0
 - Add Priority queues